SDF Meshing now tries some kind of adaptive LOD using octrees. In progress + messy
The octree is built on unique_ptr and is therefore trash. Todo - switch to vector/node indexing (we already have indexing ffs!)
We now only remesh/refine nodes that are onscreen. Todo - is this actually a good idea? It guarantees you always see the worst artifacts?!
Octree max depth should be based on smallest node dimensions, NOT a fixed depth
We could (and should...eventually...) make one giant vb for the entire octree and share it across meshes (this would be a huge perf win, but even more complexity. also major render change?)